Minor cleanup and reorganization around diagnostics
