+++ src/Inciter/NodeDiagnostics.h
@@ -14,19 +14,9 @@
 
 #include "Discretization.h"
 #include "PUPUtil.h"
+#include "Diagnostics.h"
 
 namespace inciter {
-
-//! Number of entries in diagnostics vector (of vectors), see compute()
-const std::size_t NUMDIAG = 6;
-
-//! NodeDiagnostics labels
-enum Diag { L2SOL=0,    //!< L2 norm of numerical solution
-            L2ERR,      //!< L2 norm of numerical-analytic solution
-            LINFERR,    //!< L_inf norm of numerical-analytic solution
-            ITER,       //!< Iteration count
-            TIME,       //!< Physical time
-            DT };       //!< Time step size
 
 //! NodeDiagnostics class used to compute diagnostics while integrating PDEs
 class NodeDiagnostics {
